% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.flux.stor.R
\name{def.flux.stor}
\alias{def.flux.stor}
\title{Definition function: calculate ecse dp04}
\usage{
def.flux.stor(dataInp, lvlTowr)
}
\arguments{
\item{\code{dataInp}}{Input data.}

\item{\code{lvlTowr}}{the tower levels}
}
\value{
\code{rpt} is list returned that consists of the intepolated data.
}
\description{
Definition function. Function calculates ecse dp04
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
Ke Xu \email{kxu@battelleecology.org}
}
\keyword{ECSE,}
\keyword{dp04,}
\keyword{flux}
\keyword{storage}
