% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.itpl.spce.R
\name{def.itpl.spce}
\alias{def.itpl.spce}
\title{Definition function: spatial intepolation of data at discontinous levels into spatially continous data}
\usage{
def.itpl.spce(dataInp, methItpl = "linear", resoSpceOut, lvlTowr)
}
\arguments{
\item{\code{dataInp}}{Input data.}

\item{\code{methItpl}}{intepolation method: "linear" or "constant", linear intepolation is the default method. Reference for constant spatial intepolation: Montagnani et al., 2018, Estimating the storage term in eddy covariance measurements: the ICOS methodology}

\item{\code{resoSpceOut}}{the output spatial resolution}

\item{\code{lvlTowr}}{the tower levels}
}
\value{
\code{rpt} is list returned that consists of the intepolated data.
}
\description{
Definition function. Function spatial intepolation of data at discontinous levels into spatially continous data
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
Ke Xu \email{kxu@battelleecology.org}
}
\keyword{ECSE}
\keyword{intepolate,}
\keyword{space,}
\keyword{spatial,}
