% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.itpl.time.R
\name{def.itpl.time}
\alias{def.itpl.time}
\title{Definition function: temporal intepolation from discontinous data points into temporally continous data}
\usage{
def.itpl.time(dataInp, methItpl, gapMax)
}
\arguments{
\item{\code{dataInp}}{Input data.}

\item{\code{methItpl}}{intepolation method: linear intepolation is the default method.}

\item{\code{resoTimeOut}}{the output temporal resolution}

\item{\code{gapMax}}{maximum gap for intepolation. Gaps exist this parameter won't be intepolated due to long missing data [s].}
}
\value{
\code{rpt} is list returned that consists of the intepolated data.
}
\description{
Definition function. Function temporal intepolation from discontinous data points into temporally continous data
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
Ke Xu \email{kxu@battelleecology.org}
}
\keyword{ECSE}
\keyword{intepolate,}
\keyword{time,}
