% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.shft.time.isoCo2.R
\name{def.shft.time.isoCo2}
\alias{def.shft.time.isoCo2}
\title{Definition function: Identify the time offset and adjust the crdCo2 data accordingly.}
\usage{
def.shft.time.isoCo2(
  dataList,
  qfqmList,
  site,
  lvls,
  test = FALSE,
  ofstTest = NA
)
}
\arguments{
\item{\code{dataList}}{Input data.}

\item{\code{qfqmList}}{Input quality flags.}

\item{\code{site}}{Site location in character format.}

\item{\code{lvls}}{Number of measurement level.}

\item{\code{test}}{A logical stating if this is a test. Defaults to FALSE.}

\item{\code{ofstTest}}{A user provides a time offset in seconds when test sets to TRUE.}
}
\value{
\code{rpt} is list returned that consists of the  corrected timestamps of data and qfqm and time offset in seconds.
}
\description{
Definition function. Identify the time offset and adjust the crdCo2 data accordingly.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Rich Fiorella  \email{rich.fiorella@utah.edu}
Chris Florian  \email{cflorian@battelleecology.org}
Natchaya Pingintha-Durden  \email{ndurden@battelleecology.org}
}
\keyword{Picarro,}
\keyword{crdCo2}
\keyword{offset,}
\keyword{time}
