% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.time.rate.diff.R
\name{def.time.rate.diff}
\alias{def.time.rate.diff}
\title{Definition function: calculate time rate of change for ecse dp02 data and qfqm}
\usage{
def.time.rate.diff(dataInp, numDate, PrdWndwAgr, PrdIncrAgr, Date,
  qfqmFlag = FALSE)
}
\arguments{
\item{\code{dataInp}}{Input data.}

\item{\code{numDate}}{number of dates of the input data}

\item{\code{PrdWndwAgr}}{window period, 240 s as the default}

\item{\code{PrdIncrAgr}}{incremental period, 30 min, 1800 s as the default}

\item{\code{Date}}{Date of the dataInp}

\item{\code{qfqmFlag}}{whether the input data is actual data or qfqm, FALSE as default}
}
\value{
\code{rpt} is list returned that consists of the time rate of change of data and qfqm.
}
\description{
Definition function. Calculate time rate of change for ecse dp02
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
Ke Xu \email{kxu@battelleecology.org}
}
\keyword{ECSE}
\keyword{intepolate,}
\keyword{time,}
