% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.dp01.ecse.R
\name{wrap.dp01.ecse}
\alias{wrap.dp01.ecse}
\title{Wrapper function: Preprocessing and computing NEON eddy-covariance stroage exchange L1 data product descriptive statistics}
\usage{
wrap.dp01.ecse(
  dp01 = c("co2Stor", "h2oStor", "tempAirLvl", "tempAirTop", "isoCo2", "isoH2o",
    "ch4Conc")[1],
  lvl,
  lvlMfcSampStor = NULL,
  lvlEnvHut = NULL,
  lvlValv = NULL,
  lvlCrdH2oValvVali = NULL,
  lvlCrdCo2Valv = NULL,
  data = list(),
  qfInp = list(),
  TypeMeas = c("samp", "vali")[1],
  PrdMeas,
  PrdAgr,
  idxTime = list()
)
}
\arguments{
\item{dp01}{A vector of class "character" containing the name of NEON ECSE dp01 which descriptive statistics are being calculated, \cr
c("co2Stor", "h2oStor", "tempAirLvl", "tempAirTop", "isoCo2", "isoH2o"). Defaults to "co2Stor". [-]}

\item{lvl}{Measurement level of dp01 which descriptive statistics are being calculated. Of type character. [-]}

\item{lvlMfcSampStor}{Measurement level of mfcSampStor which apply to only  dp01 equal to "co2Stor" or "h2oStor". Defaults to NULL. Of type character. [-]}

\item{lvlEnvHut}{Measurement level of envHut. Defaults to NULL. Of type character. [-]}

\item{lvlValv}{Measurement level of irgaValvLvl, crdCo2ValvLvl, or crdH2oValvLvl. Defaults to NULL. Of type character. [-]}

\item{lvlCrdH2oValvVali}{Measurement level of crdH2oValvVali which apply to only  dp01 equal to "isoH2o". Defaults to NULL. Of type character. [-]}

\item{lvlCrdCo2Valv}{Horizontal and vertical location for crdCo2 valve. Defaults to NULL. Of type character. [-]}

\item{data}{A list of data frame containing the input dp0p data that related to dp01 which descriptive statistics are being calculated. Of class integer". [User defined]}

\item{qfInp}{A list of data frame containing the input quality flag data that related to dp01 are being grouped. Of class integer". [-]}

\item{TypeMeas}{A vector of class "character" containing the name of measurement type (sampling or validation), TypeMeas = c("samp", "vali"). Defaults to "samp". [-]}

\item{PrdMeas}{The measurement time period in minute.  [min]}

\item{PrdAgr}{The time period to aggregate to averaging in minute. [min]}

\item{idxTime}{A list of data frame containing the indices and corresponding times for aggregation periods. [-]}
}
\value{
A list of dp01 descriptive statistics. \cr 
\code{mean} The mean of non-NA values in \code{data}
\code{min} The minimum value of non-NA values in \code{data}
\code{max} The maximum value of non-NA values in \code{data}
\code{vari} The variance of non-NA values in \code{data}
\code{numSamp} The number of non-NA values in \code{data}
\code{se} The standard error of the mean of non-NA values in \code{data}
\code{timeBgn} The beginning time to determine descriptive statistics.
\code{timeEnd} The ending time to determine descriptive statistics.
}
\description{
Wrapper function. Preprocessing and computing NEON eddy-covariance stroage exchange (ECSE) Level 1 data product (dp01) descriptive statistics (mean, minimum, maximum, variance, number of non-NA points).
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Natchaya Pingintha-Durden \email{eddy4R.info@gmail.com}
}
\keyword{ECSE}
\keyword{aggregate,}
\keyword{average,}
\keyword{descriptive}
\keyword{dp01,}
\keyword{statistics,}
