% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.shft.time.isoH2o.R
\name{def.shft.time.isoH2o}
\alias{def.shft.time.isoH2o}
\title{Definition function: Identify the time offset and adjust the crdH2o data accordingly.}
\usage{
def.shft.time.isoH2o(dataList, qfqmList, valvVali, site, lvls)
}
\arguments{
\item{\code{dataList}}{Input data.}

\item{\code{qfqmList}}{Input quality flags.}

\item{\code{site}}{Site location in character format.}

\item{\code{lvls}}{Number of measurement level.}
}
\value{
\code{rpt} is list returned that consists of the  corrected timestamps of data and qfqm and time offset in seconds.
}
\description{
Definition function. Identify the time offset and adjust the crdH2o data accordingly.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden  \email{ndurden@battelleecology.org}
Rich Fiorella  \email{rich.fiorella@utah.edu}
Chris Florian  \email{cflorian@battelleecology.org}
}
\keyword{Picarro,}
\keyword{crdH2o}
\keyword{offset,}
\keyword{time}
